Completes day 01
This is the initial commit for day 1. It could be refined a lot. There
are several extra procedures I used while developing.

The part I was most pleased with was #:conversion-map, which allowed me
to use the code from part 1 and add the functionality for part 2 based
on the value of the keyword argument used in #:conversion-map. This gave
some justification to the software design study I've been doing lately.
It was an idea I remembered seeing (not sure if I remembered before or
after I thought of it) from Paul Graham (I think but can't find the
exact article now) who used optional parameters to add new functionality
to running code without distrubing the already-running code.

I think this part of the puzzle was incomplete or inconsistent. When
moving from right to left "2one" returns "1" as the last number but when
converting from right to left "twone" returns "2" because the conversion
is done from left to right and the the number pulled from right to left.
I think the puzzle would have been more consistent in "twone" was
converted from right to left as well as being pulled from right to left,
resulting in a conversion to "tw1" and a right-most number of "1".