Completes day 02
Roughly the finished version of day 2. Most of the effort was in the
initial parsing of the game line.

The design was easier due to the struct. As I thought through how to
break down the line I realized I needed a data structure that was
simple, easy to create, and allowed later comparison.

Also by God's grace I realized that parsing the throws/draws out of the
bag by filtering them for the max value of each color would help looping
over them later. It happened to be the case that other throws/draws from
the bag were not needed, which was a relief. The initial parser would
have changed for part two if other information from the throws/draws was
needed. This is a great reminder of the necessity of taking time to
think through the data structure most appropriate to the problem.